Rename DragonTCP.h to DragonTCP.hpp and update CMakeLists.txt in all folders. Update CMakeLists.txt to install in DragonTCP directories and make sure to install the new header file (it was renamed).
