+++ src/tcp.cpp
@@ -16,7 +16,7 @@
 * along with lidar project.  If not, see https://www.gnu.org/licenses/ 
 */ 
 
-#include "DragonTCP.h"
+#include "DragonTCP.hpp"
 
 namespace asio = boost::asio;
 using asio::ip::tcp;
